Remove parser corpus tests. Yay!
