--- parser/src/tests/corpus.rs
@@ -1,170 +1,0 @@
-mod error;
-
-use self::error::{Error, Errors, Mismatch, Reason};
-use include_dir::{include_dir, Dir, File};
-use sappho_ast::PureExpr;
-use std::path::Path;
-
-static CORPUS_DIR: Dir<'_> = include_dir!("$CARGO_MANIFEST_DIR/src/tests/corpus");
-
-#[test]
-fn positives() {
-    parse_corpus("positives", parse_file)
-}
-
-#[test]
-fn negatives() {
-    parse_corpus("negatives", parse_file_negative)
-}
-
-fn parse_corpus<F, T>(corpusname: &str, parsefunc: F)
-where
-    F: Fn(&'static Path, &'static str) -> Result<T, Reason>,
-    T: ToString,
-{
-    if let Some(e) = parse_corpus_result(corpusname, parsefunc).err() {
-        panic!("{}", e);
-    }
-}
-
-fn parse_corpus_result<F, T>(corpusname: &str, parsefunc: F) -> Result<(), Errors>
-where
-    F: Fn(&'static Path, &'static str) -> Result<T, Reason>,
-    T: ToString,
-{
-    let mut errors = Errors::default();
-
-    let corpcase = CORPUS_DIR
-        .get_dir(corpusname)
-        .unwrap_or_else(|| panic!("src/tests/corpus/{} not found", corpusname));
-
-    for casedir in only_dirs(corpcase) {
-        if let Some(suberrs) = parse_case(casedir, &parsefunc).err() {
-            errors.extend(suberrs);
-        }
-    }
-
-    errors.into_result()
-}
-
-fn only_dirs<'a>(d: &Dir<'a>) -> Vec<&'a Dir<'a>> {
-    use include_dir::DirEntry::Dir;
-
-    let mut ds = vec![];
-    for entry in d.entries() {
-        match entry {
-            Dir(d) => ds.push(d),
-            _ => panic!("Unexpected file: {:#?}", entry),
-        }
-    }
-    ds
-}
-
-fn parse_case<F, T>(casedir: &'static Dir, parsefunc: F) -> Result<(), Errors>
-where
-    F: Fn(&'static Path, &'static str) -> Result<T, Reason>,
-    T: ToString,
-{
-    let expected = file_contents(casedir, "expected")?.trim_end();
-    let inputs = get_testcase_inputs(casedir)?;
-
-    inputs
-        .into_iter()
-        .map(|f| parse_case_input(f, expected, &parsefunc))
-        .collect::<Errors>()
-        .into_result()
-}
-
-fn get_testcase_inputs<'a>(casedir: &'a Dir<'a>) -> Result<Vec<&'a File<'a>>, Error> {
-    let mut inputs: Vec<&File> = vec![];
-    for f in casedir.files() {
-        let fname = f
-            .path()
-            .file_name()
-            .and_then(|os| os.to_str())
-            .ok_or(Reason::BadPath)
-            .map_err(|r| Error(f.path().to_path_buf(), r))?;
-
-        if fname.starts_with("input") {
-            inputs.push(f);
-        } else if fname != "expected" {
-            return Err(Error(f.path().to_path_buf(), Reason::UnexpectedFile));
-        }
-    }
-
-    if inputs.is_empty() {
-        Err(Error(
-            casedir.path().to_path_buf(),
-            Reason::MissingFile("<no 'input*' files for case directory>"),
-        ))
-    } else {
-        Ok(inputs)
-    }
-}
-
-fn parse_case_input<F, T>(
-    inputfile: &'static File,
-    expected: &str,
-    parsefunc: F,
-) -> Result<(), Error>
-where
-    F: Fn(&'static Path, &'static str) -> Result<T, Reason>,
-    T: ToString,
-{
-    parse_case_input_reason(inputfile, expected, &parsefunc)
-        .map_err(|r| Error(inputfile.path().to_path_buf(), r))
-}
-
-fn parse_case_input_reason<F, T>(
-    inputfile: &'static File,
-    expected: &str,
-    parsefunc: F,
-) -> Result<(), Reason>
-where
-    F: Fn(&'static Path, &'static str) -> Result<T, Reason>,
-    T: ToString,
-{
-    let input: &'static str = std::str::from_utf8(inputfile.contents())?;
-    let inpath = inputfile.path();
-
-    match parsefunc(inpath, input).map(|v| v.to_string()) {
-        Ok(found) if found.trim_end() == expected => Ok(()),
-
-        // Allow a missing "expected" file as a dev convenience:
-        Ok(found) => Err(Reason::MismatchedOutput(Mismatch {
-            found: found.trim_end().to_string(),
-            expected: expected.to_string(),
-        })),
-
-        Err(reason) => Err(reason),
-    }
-}
-
-fn file_contents<'a>(d: &'a Dir, fname: &'static str) -> Result<&'a str, Error> {
-    file_contents_to_reason(d, fname).map_err(|r| Error(d.path().join(fname), r))
-}
-
-fn file_contents_to_reason<'a>(d: &'a Dir, fname: &'static str) -> Result<&'a str, Reason> {
-    let bytes = d
-        .get_file(d.path().join(fname))
-        .map(|f| f.contents())
-        .unwrap_or(b"");
-    let src = std::str::from_utf8(bytes)?;
-    Ok(src)
-}
-
-fn parse_file(path: &'static Path, source: &'static str) -> Result<PureExpr, Reason> {
-    let expr = crate::parse((path, source))?;
-    Ok(expr)
-}
-
-fn parse_file_negative(
-    path: &'static Path,
-    source: &'static str,
-) -> Result<crate::LoadParseError<'static>, Reason> {
-    match parse_file(path, source) {
-        Ok(expr) => Err(Reason::InvalidParse(expr)),
-        Err(Reason::Parse(e)) => Ok(e),
-        Err(e) => Err(e),
-    }
-}
